requirements.txt
requests
pytube==15.0.0
transformers==4.32.1
python-dotenv==0.21.0
keybert==0.8.4
keyphrase_vectorizers
torch==2.2.0
Pillow==9.4.0